% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{constraint}
\alias{constraint}
\title{Burt's Constraint Index.}
\usage{
constraint(g, v = igraph::V(g))
}
\arguments{
\item{g}{The igraph object to analyze.}

\item{v}{vertices over which to compute constraint (default to all)}
}
\value{
A numeric vector with the constraint score for each vertex in v
}
\description{
The igraph package provides an implementation of Constraint; this is an alternate implementation.
}
\examples{
ig.ex <- igraph::erdos.renyi.game(100, p.or.m=0.3) # generate an undirected 'igraph' object
constraint(ig.ex) # constraint scores for each node in the graph
}

