% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{csv.to.igraph}
\alias{csv.to.igraph}
\title{Convert a CSV file to an igraph graph object.}
\usage{
csv.to.igraph(fname)
}
\arguments{
\item{fname}{A filename}
}
\value{
An igraph graph object built from the filename.
}
\description{
The first column should be sources, the second should be targets.
}

