% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{dimacs.to.graph}
\alias{dimacs.to.graph}
\title{Convert a DIMACS file to an igraph graph object.}
\usage{
dimacs.to.graph(fname)
}
\arguments{
\item{fname}{A filename}
}
\value{
An igraph graph object built from the filename.
}
\details{
DIMACS is a very simple ASCII format for graphs. The first line is of the
form:

p sp <n> <m>

where <n> is the number of nodes in the graph and <m> is the number of edges.
Each subsequent line represents an edge, and is of the form:

e <u> <v> <weight>

where <u> and <v> are vertex IDs in the range 1-n, and weight is the weight
of the edge (or 1 if edges are unweighted).
}
\examples{
dimacs.to.graph("mygraph.dim")
}

