% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{eigencentrality}
\alias{eigencentrality}
\title{Eigencentrality centrality measure.}
\usage{
eigencentrality(g)
}
\arguments{
\item{g}{The igraph object to analyze}
}
\value{
A numeric vector with the eigencentrality score for each vertex
}
\description{
Eigencentrality is a measure of the influence of a node in a network based
on the eigendecompositon of the adjacency matrix. This is a thin wrapper
around the evcent function in the igraph package.
}

