% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{ergm.to.igraph}
\alias{ergm.to.igraph}
\title{Convert an ergm network object to an igraph object.}
\usage{
ergm.to.igraph(net)
}
\arguments{
\item{net}{A network object}
}
\value{
An igraph object
}
\description{
Convert an ergm network object to an igraph object.
}

