% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{igraph.to.network}
\alias{igraph.to.network}
\title{Convert an igraph graph object to a network object (from the "network"
package).}
\usage{
igraph.to.network(g)
}
\arguments{
\item{g}{An igraph object}
}
\value{
A network object
}
\description{
Convert an igraph graph object to a network object (from the "network"
package).
}

