% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influencer-package.R
\docType{package}
\name{influencer-package}
\alias{influencer}
\alias{influencer-package}
\title{influencer: R Tools for Connectome Influence Analysis}
\description{
Tools for calculating influence scores of neurons in connectomes using linear dynamical models. Provides both R implementations and Python wrappers for the ConnectomeInfluenceCalculator library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://natverse.github.io/influencer}
  \item \url{https://github.com/natverse/influencer}
  \item Report bugs at \url{https://github.com/natverse/influencer/issues}
}

}
\author{
\strong{Maintainer}: Alexander Bates \email{alexander.shakeel.bates@gmail.com} (\href{https://orcid.org/0000-0002-1195-0445}{ORCID})

}
\keyword{internal}
