% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{betweenness}
\alias{betweenness}
\title{Compute the betweenness values.}
\usage{
betweenness(g, snap = T)
}
\arguments{
\item{g}{The igraph object to analyze}

\item{snap}{True to use the SNAP betweenness code, False to use igraph::betweenness}
}
\value{
A numeric vector with the betweenness centrality score for each vertex
}
\description{
Compute the betweenness values.
}

