% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{bridging}
\alias{bridging}
\title{Valente's bridging vertex measure.}
\usage{
bridging(g)
}
\arguments{
\item{g}{The igraph object to analyze.}
}
\value{
A numeric vector with the bridging score for each vertex
}
\description{
A node's bridging score is the average decrease in cohesiveness if each of
its edges were removed from the graph.
}
\seealso{
\url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC2889704/}
}

