% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{bridging}
\alias{bridging}
\title{Compute bridging scores for a given graph.}
\usage{
bridging(g, MPI = F)
}
\arguments{
\item{g}{The igraph object to analyze.}

\item{MPI}{True to use MPI}
}
\description{
Compute bridging scores for a given graph.
}

