% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{bridging}
\alias{bridging}
\title{Compute bridging scores for a given graph.}
\usage{
bridging(g, MPI = F)
}
\arguments{
\item{g}{The igraph object to analyze.}

\item{MPI}{True to use MPI}
}
\value{
A numeric vector with the bridging score for each vertex

This function can use MPI in order to parellelize the bridging computation
across vertices. If you would like to use MPI:
\itemize{
\item{}{An MPI implementation, such as OpenMPI, must be installed on your system}
\item{}{R must be compiled with the --with-mpi option}
\item{}{the Rmpi and snow packages must be installed and loaded prior to the bridging invocation}
\item{}{start R with the "mpirun -n 1" wrapper}
\item{}{set MPI=True in the arguments to bridging}
}
}
\description{
A node's bridging score is the average decrease in cohesiveness if each of
its edges were removed from the graph.
}
\examples{
\dontrun{
library(Rmpi)
library(snow)
library(influenceR)
g <- dimacs.to.graph("mygraph.dim")
bridging(g, MPI=T)
}
}
\seealso{
\url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC2889704/}
}

