% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{constraint}
\alias{constraint}
\title{Burt's graph constraint vertex measure.}
\usage{
constraint(g, v = igraph::V(g))
}
\arguments{
\item{g}{The igraph object to analyze.}

\item{v}{vertices over which to compute constraint (default to all)}
}
\value{
A numeric vector with the constraint score for each vertex in v
}
\description{
This is an alternative to the implementation of Burt's consraint in the
igraph package.
}

