% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{ens}
\alias{ens}
\title{Burt's effective network size vertex measure.}
\usage{
ens(g)
}
\arguments{
\item{g}{The igraph object to analyze.}
}
\value{
A numeric vector with the effective network size for each vertex
}
\description{
Burt's effective network size vertex measure.
}

