% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{keyplayer}
\alias{keyplayer}
\title{Compute a KPP-Neg set for a given graph.}
\usage{
keyplayer(g, k, prob = 0, tol = 1e-04, maxsec = 600, roundsec = 30)
}
\arguments{
\item{g}{The igraph object to analyze.}

\item{k}{The size of the KP-set}

\item{prob}{probability of accepting a state with a lower value}

\item{tol}{tolerance within which to stop the optimization and accept the current value}

\item{maxsec}{The total computation budget for the optimization, in seconds}

\item{roundsec}{Number of seconds in between synchronizing workers' answer}
}
\description{
Compute a KPP-Neg set for a given graph.
}

