% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{keyplayer}
\alias{keyplayer}
\title{Compute a KPP-Neg set for a given graph.}
\usage{
keyplayer(g, k, prob = 0, tol = 1e-04, maxsec = 600, roundsec = 30)
}
\arguments{
\item{g}{The igraph object to analyze.}

\item{k}{The size of the KP-set}

\item{prob}{probability of accepting a state with a lower value}

\item{tol}{tolerance within which to stop the optimization and accept the current value}

\item{maxsec}{The total computation budget for the optimization, in seconds}

\item{roundsec}{Number of seconds in between synchronizing workers' answer}
}
\value{
a vector with the vertex number of each vertex in the selected set S.
}
\description{
The "Key Player" family of node importance algorithms (Borgatti 2006) involves the selection
of a metric of node importance and a combinatorial optimization strategy to
choose the set S of vertices of size k that maximize that metric. This
algorithm uses the KPP-Neg metric, whereby we sum over all vertices not in S the reciprocal
of the shortest distance to a vertex in S. For combinatorial optimization, we use
stochastic gradient descent, where in each optimization round, we select a node u in S
and v not in S at random, switch them, and accept the switch if evaluation of the
metric improves. This implementation uses OpenMP so multiple workers can explore the
solution space in parallel, synchronizing to pick the best answer after a given computation
budget has elapsed.
}

